+++ src/components/split-button/split-button.js
@@ -109,7 +109,12 @@ class SplitButton extends React.Component {
    * @return {String} Main className
    */
   get additionalButtonsClasses() {
-    return 'carbon-split-button__additional-buttons';
+    return classNames(
+      'carbon-split-button__additional-buttons',
+      {
+        'carbon-split-button__additional-buttons--hidden': !this.state.showAdditionalButtons
+      }
+    );
   }
 
   /**
@@ -198,7 +203,7 @@ class SplitButton extends React.Component {
     return (
       <div className={ this.mainClasses } onMouseLeave={ this.hideButtons }>
         { this.renderMainButton }
-        { this.state.showAdditionalButtons ? this.renderAdditionalButtons : null}
+        { this.renderAdditionalButtons }
       </div>
     );
   }