+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * Add `additionalRequestParams` prop to `DropdownFilterAjax`
 
+# 0.35.2
+
+* Hide SplitButton additional buttons instead of removing them.
+
 # 0.35.1
 
 * Ensure that node modules can only upgrade patch versions